+++ tests/Fixture/PureEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ByCerfrance\JsonFragments\Tests\Fixture;
+
+enum PureEnum
+{
+    case Ready;
+}