+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.0.1] - 2026-09-25
+
+### Fixed
+
+- Accept backed enums as JSON values, preserving `JsonSerializable` precedence
+  and validating their scalar backing values.
+
 ## [1.0.0] - 2026-09-24
 
 - Initial release.